+++ lib/src/widgets/default_splash_screen.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+const defaultSplashScreen = Text(
+  'Splash Screen',
+  style: TextStyle(
+    color: Color(0xff71C6D1),
+    fontWeight: FontWeight.w800,
+    fontSize: 32,
+  ),
+);